move prcheck logic to utils package
